oh god iforgottoremovetheturbolinks
removes turbolinks. phew.

In other news, rearrange the JS and start using translations to remind myself later on that this might be needed in production later. After all, we might need a convenient way to manage all the strings and I do _not_ wish to do that, if there is qualified people who are some much better with the words.